requirements.txt
boto3==1.34.36
botocore==1.34.36
jmespath==1.0.1
pillow==10.2.0
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
